requirements.txt
tkinter
Pillow
